Shift views form main app to product app
